Area headings
Added the projects, announcements and trending
headings. I've had to use relative positioning to
get them to the right spot. For announcements and
trending they have - bottom margin to allow for
the content to spread nicely. The projects title
is actually part of the header, and relatively
positioned lower because putting it inside the
projects div made it take up a slot in the grid
and it would push down the cards.